Refactor unit tests into logical units.
